+++ examples/fastify-http.ts
@@ -1,3 +1,6 @@
+/**
+ * This example shows how to use this library with Fastify using HTTP
+ */
 
 import { dirname, join } from 'path';
 import { fileURLToPath } from 'url';