+++ examples/pre-compressed-util.ts
@@ -1,20 +1,15 @@
+/**
+ * This file is an example of file precompressing to use with ./pre-compressed.ts example
+ */
+
 import fs from 'fs';
-import zlib from 'zlib';
-import stream from 'stream';
 import { join, dirname } from 'path';
+import { pipeline } from 'stream';
+import { promisify } from 'util';
+import zlib from 'zlib';
 import { fileURLToPath } from 'url';
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function handleEvent(resolve: () => void, reject: (err: any) => void) {
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	return (err: any) => {
-		if (err) {
-			reject(err);
-		} else {
-			resolve();
-		}
-	};
-}
+const promisifiedStreamPipeline = promisify(pipeline);
 
 async function precompressWithGzipAndBrotli(directoryPath: string) {
 	const directoryFiles = await fs.promises.readdir(directoryPath, { withFileTypes: true });
@@ -35,30 +30,19 @@ async function precompressWithGzipAndBrotli(directoryPath: string) {
 		) {
 			return;
 		}
-		const fileContents = fs.createReadStream(filepath);
-		const zip = zlib.createGzip({
+		const gzip = zlib.createGzip({
 			level: zlib.constants.Z_BEST_COMPRESSION,
 		});
-		const gzipCopy = new stream.PassThrough();
-		fileContents.pipe(gzipCopy);
 		const gzWriteStream = fs.createWriteStream(`${ filepath }.gz`);
 		const br = zlib.createBrotliCompress({ params: {
 			[zlib.constants.BROTLI_PARAM_MODE]: zlib.constants.BROTLI_MODE_TEXT,
 			[zlib.constants.BROTLI_PARAM_SIZE_HINT]: stat.size,
 			[zlib.constants.BROTLI_PARAM_QUALITY]: zlib.constants.BROTLI_MAX_QUALITY,
 		} });
-		const brCopy = new stream.PassThrough();
-		fileContents.pipe(brCopy);
 		const brWriteStream = fs.createWriteStream(`${ filepath }.br`);
 		await Promise.all([
-			new Promise<void>((resolve, reject) => {
-				gzipCopy.pipe(zip).pipe(gzWriteStream)
-					.on('finish', handleEvent(resolve, reject));
-			}),
-			new Promise<void>((resolve, reject) => {
-				brCopy.pipe(br).pipe(brWriteStream)
-					.on('finish', handleEvent(resolve, reject));
-			}),
+			promisifiedStreamPipeline(fs.createReadStream(filepath), gzip, gzWriteStream),
+			promisifiedStreamPipeline(fs.createReadStream(filepath), br, brWriteStream),
 		]);
 	}));
 }
@@ -68,5 +52,5 @@ precompressWithGzipAndBrotli(join(dirname(fileURLToPath(import.meta.url)), 'asse
 		console.info('files have been precompressed');
 	})
 	.catch(err => {
-		console.error(err);
+		console.error('error:', err);
 	});